Detect and present frames at constant and variable rates
Three presentation modes are added:
 - CFR: constant frame rate
 - VFR: variable frame rate
 - LIMIT_VFR: limited variable frame

For LIMIT_VFR, the rate runs no faster than the desktop
refresh rate, which is now added as a member variable.
to the sdl.desktop struct.

The added DeterminePresentationMode function uses simple
herustics to determine which of the above presentation
modes should be used for a given system.  These could be
added as a (yet another) conf options, however the
herustics have proven to make the right choice given
many different test systems (24 Hz frame-buffer, 60 Hz
1080p, and 144 Hz freesync).

The added function IsFrameAllowed is used to determine
when frames are allowed to be shown, given a desire period.

The added GFX_MaybePresentFrame function takes into account
the given presentation mode and uses IsFrameAllowed
(both in the CFR and LIMIT_VFR cases) to determine when to
present a frame.